+++ lib/model/db/user.dart
@@ -11,6 +11,7 @@ class UserTypes {
   static const String OPENCAP = "opencap";
   static const String CONTACT = "contact";
   static const String ONCHAIN = "onchain";
+  static const String WELL_KNOWN = "well_known";
 }
 
 String? lowerStripAddress(String? address) {
@@ -126,10 +127,10 @@ class User {
   }
 
   static String? getDisplayNameWithType(String? name, String? userType) {
-    if (userType == UserTypes.ONCHAIN) {
+    if (userType == UserTypes.ONCHAIN || userType == UserTypes.WELL_KNOWN) {
       return "@${name!}";
     } else if (userType == UserTypes.NANO_TO) {
-      return "@${name!}";
+      return "${name!}@nano.to";
     } else {
       return name;
     }