+++ lib/ui/receive/receive_sheet.dart
@@ -24,6 +24,7 @@ import 'package:wallet_flutter/ui/receive/receive_show_qr.dart';
 import 'package:wallet_flutter/ui/receive/split_bill_sheet.dart';
 import 'package:wallet_flutter/ui/request/request_confirm_sheet.dart';
 import 'package:wallet_flutter/ui/send/send_sheet.dart';
+import 'package:wallet_flutter/ui/util/confirm_sheet.dart';
 import 'package:wallet_flutter/ui/util/formatters.dart';
 import 'package:wallet_flutter/ui/util/handlebars.dart';
 import 'package:wallet_flutter/ui/util/routes.dart';
@@ -177,7 +178,7 @@ class _ReceiveSheetState extends State<ReceiveSheet> {
             _addressController.text.length > 1 &&
             SendSheetHelpers.isSpecialAddress(_addressController.text)) {
           final String formattedAddress = SendSheetHelpers.stripPrefixes(_addressController.text);
-          if (_addressController.text != formattedAddress) {
+          if (_addressController.text != formattedAddress && !SendSheetHelpers.isWellKnown(_addressController.text)) {
             setState(() {
               _addressController.text = formattedAddress;
             });
@@ -215,7 +216,18 @@ class _ReceiveSheetState extends State<ReceiveSheet> {
         // check if UD / ENS / opencap / onchain address:
         if (_addressController.text.isNotEmpty && !_addressController.text.contains("★")) {
           User? user = await sl.get<DBHelper>().getUserOrContactWithName(_addressController.text);
-          user ??= await sl.get<UsernameService>().figureOutUsernameType(_addressController.text);
+          if (user == null) {
+            if (!mounted) return;
+            final bool? confirmed = await Sheets.showAppHeightSmallSheet(
+              context: context,
+              widget: ConfirmSheet(subtitle: Z.of(context).checkUsernameConfirmInfo),
+              allowSlide: true,
+            ) as bool?;
+            
+            if (confirmed == true) {
+              user ??= await sl.get<UsernameService>().figureOutUsernameType(_addressController.text);
+            }
+          }
 
           if (user != null) {
             setState(() {